# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD API",
        "version": "1.0"
    },
    "paths": {
        "/api/pets": {
            "get": {
                "tags": [
                    "Pet"
                ],
                "summary": "get pet list",
                "operationId": "ddd551168b1d5b280be068f61a15b92d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PetResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pet"
                ],
                "summary": "store pet",
                "operationId": "5bbbd00f35cb6067370e16b90bc17ecb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "pet name"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get user list",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "store user",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "yuta"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "get user",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "update user",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "yuta"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "test@test.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "delete user",
                "operationId": "5a702a3f11c33d10743e96488ba9be06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/path": {
            "get": {
                "operationId": "f8672c2562b04fa0ad13dedd825d61c7",
                "responses": {
                    "200": {
                        "description": "An example path"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PetResource": {
                "properties": {
                    "id": {
                        "description": "pet id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "pet name",
                        "type": "string",
                        "example": "pet"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "name": {
                        "description": "user name",
                        "type": "string",
                        "example": "yuta"
                    },
                    "email": {
                        "description": "user email",
                        "type": "email",
                        "example": "test@test.com"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "path": {
                        "type": "string",
                        "example": "http://example.com/pagination"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "path": {
                        "type": "string",
                        "example": "http://example.com/pagination"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "PaginationResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {}
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "description": "現在のページ",
                "example": 1
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "１ページあたりに表示する件数",
                "example": 1
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}